Fixed issue #371 Replace the implementation=file-system with data-store=file-system + package renaming
