+++ strongbox-storage/strongbox-storage-core/src/main/java/org/carlspring/strongbox/providers/datastore/StorageProvider.java
@@ -1,4 +1,4 @@
-package org.carlspring.strongbox.providers.storage;
+package org.carlspring.strongbox.providers.datastore;
 
 import java.nio.file.FileSystem;
 import java.nio.file.spi.FileSystemProvider;
