+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/BaseArtifactController.java
@@ -1,7 +1,7 @@
 package org.carlspring.strongbox.controllers;
 
 import org.carlspring.strongbox.providers.layout.LayoutProviderRegistry;
-import org.carlspring.strongbox.providers.storage.StorageProviderRegistry;
+import org.carlspring.strongbox.providers.datastore.StorageProviderRegistry;
 import org.carlspring.strongbox.storage.Storage;
 import org.carlspring.strongbox.storage.repository.Repository;
